[refactor] Test용 API 체크스타일 수정
