+++ build.gradle
@@ -21,6 +21,7 @@ def coverageExcludePackages = [
     '**/dto/**',         // DTO 클래스 제외
     '**/apipayload/**',  // API 공통 응답 및 예외 처리 제외
     '**/security/**', // 시큐리티 관련 클래스 제외
+    '**/apitest/**' // 테스트용 토큰 방법 api 제외
 ]
 
 // JaCoCo용 제외 패턴 (클래스 파일)
